added the chart route
