+++ public/js/topSpendingChart.js
@@ -49,3 +49,4 @@ const createChart = async (userId) => {
     console.log(err);
   }
 };
+
