+++ views/budget.handlebars
@@ -38,7 +38,6 @@
              </div>
 
     </section>
-    <script type="module" src="/js/renderData.js"></script>
     <script src="/dist/flowbite.min.js"></script>
   </body>
   </html>
